fixed upgrade to check for docker

+++ pve_lxc_scripts/maintenance/upgrade.sh
@@ -22,11 +22,13 @@ fullupdate()
 dockerupdateall()
 {
     if [ ! -x "$(command -v pveversion)" ]; then #ensure it's not the server 
-      cd /mnt/pve-sata-ssd/ssd-data/dockerapps || exit
-      cd "$(hostname)" || exit
-      find . -maxdepth 1 -type d \( ! -name . \) -not -path '*disabled*' -exec bash -c "cd '{}' && pwd && docker compose down && docker compose pull && docker compose up -d --remove-orphans" \;
-      docker image prune -a -f
-      docker system prune --volumes -f
+      if [ -x "$(command -v docker)" ]; then #ensure docker is installed on the host 
+        cd /mnt/pve-sata-ssd/ssd-data/dockerapps || exit
+        cd "$(hostname)" || exit
+        find . -maxdepth 1 -type d \( ! -name . \) -not -path '*disabled*' -exec bash -c "cd '{}' && pwd && docker compose down && docker compose pull && docker compose up -d --remove-orphans" \;
+        docker image prune -a -f
+        docker system prune --volumes -f
+     fi
     fi
 }
